调整聊天回复目标消息id变量名

+++ src/main/java/com/jihai/bitfree/dao/MessageDAO.java
@@ -10,7 +10,7 @@ public interface MessageDAO {
 
     Integer count();
 
-    Long insert(MessageDO messageDO);
+    Integer insert(MessageDO messageDO);
 
     Integer getRecentMessageCount();
